Change from darkMode to theme
Instead of a boolean theme (dark or not), change to a generic 'theme'
setting. This keeps backwards-compatibility from older settings. The new
themes are 'dark' and 'light' and 'system'. 'system' will allow for the
user's OS to apply darkMode automatically depending on their system
settings. Null and 'system' are treated as the same. 'dark' and 'light'
will ignore OS settings.

This also introduces reactivity from the OS's changing
`prefers-color-scheme`, for example if your system is set to go into
dark mode at sunrise, then the browser will switch to
`prefers-color-scheme: dark`. Before this commit, the theme would not
react to this change until page reload. Now, it will watch for media
changes and apply the dark theme if appropriate.